Fix cumulativeUserMemory initial calculation & use double for cumulativeUserMemory consistently

+++ presto-main/src/main/java/com/facebook/presto/execution/StageExecutionStats.java
@@ -434,7 +434,7 @@ public class StageExecutionStats
                 completedDrivers,
                 rawInputDataSize,
                 rawInputPositions,
-                (long) cumulativeUserMemory,
+                cumulativeUserMemory,
                 userMemoryReservation,
                 totalMemoryReservation,
                 totalCpuTime,
